refactor: make payment uris configurable per environment
